Rename ScreenManager.h to game-source-code/ScreenManager.h
